Confirmación al borrar
Añadido confirmación al borrar de la base de datos

+++ proyecto-angular/src/app/components/mensajes-bandeja/mensajes-bandeja.component.ts
@@ -32,7 +32,7 @@ export class MensajesBandejaComponent implements OnInit {
     this.getMensajesUser(this._route.snapshot.paramMap.get('id'));
   }
 
-  ngOnInit(): void {}
+  ngOnInit(): void { }
 
   getMensajesUser(id) {
     this._mensajeService.getMensajesUser(id).subscribe(
@@ -40,14 +40,14 @@ export class MensajesBandejaComponent implements OnInit {
         this.mensajes = response.mensajes;
 
         this.mensajes.forEach(mensaje1 => {
-          let cont=0;
+          let cont = 0;
           this.mensajes.forEach(mensaje2 => {
             if (
               mensaje1.id_usuario_r === mensaje2.id_usuario_e && mensaje1.id_usuario_e === mensaje2.id_usuario_r/* &&
               this.mensajes[i].id_usuario_r == this.mensajes[j].id_usuario_e */
             ) {
               console.log("deberia de funcionar");
-              this.mensajes.splice(cont,1);
+              this.mensajes.splice(cont, 1);
             }
             cont++
           });
@@ -61,16 +61,18 @@ export class MensajesBandejaComponent implements OnInit {
     );
   }
 
-  borrarMensajes(id1, id2){
-    console.log("hola");
-    this._mensajeService.delete(this.token, id1, id2).subscribe(
-      response=>{
-        console.log("mensajes borrados");
-        this.getMensajesUser(this._route.snapshot.paramMap.get('id'));
-      },
-      error=>{
-        console.log(<any>error);
-      }
-    );
+  borrarMensajes(id1, id2) {
+    let borrar = confirm("Seguro que quiere borrar la sección?");
+    if (borrar) {
+      this._mensajeService.delete(this.token, id1, id2).subscribe(
+        response => {
+          console.log("mensajes borrados");
+          this.getMensajesUser(this._route.snapshot.paramMap.get('id'));
+        },
+        error => {
+          console.log(<any>error);
+        }
+      );
+    }
   }
 }
